bgp: T3037: migrate invalid "address-family ipv6-unicast capability dynamic" option
The mentioned CLI option is not available in FRR, there is one one generic
dynamic capability option in the root of the neighbor. This commit drops the
CLI support as it will be migrate to math FRR support.

(cherry picked from commit 0d42f7119306f2ebd6011f650be1675cb267acb7)